fixing mvc
fixing recipe controller

+++ mvc/models/reviews/reviews.go
@@ -1,9 +1,9 @@
 package reviews
 
 type Review struct {
-	ID           int    `json:"id"`
-	UserID       int    `json:"userId"`
-	RecipeID     int    `json:"recipeId"`
-	Rating       int    `json:"rating"`
-	RecipeReview string `json:"review"`
+	ID           int     `json:"id"`
+	UserID       int     `json:"userId"`
+	RecipeID     int     `json:"recipeId"`
+	Rating       float64 `json:"rating"`
+	RecipeReview string  `json:"review"`
 }
